+++ src/lib.rs
@@ -66,8 +66,8 @@ pub fn yul_to_eravm(
     libraries: Vec<String>,
     solc_path: Option<String>,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
-    is_system_mode: bool,
+    llvm_options: Vec<String>,
+    enable_eravm_extensions: bool,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
@@ -75,7 +75,7 @@ pub fn yul_to_eravm(
 
     let solc_version = match solc_path {
         Some(solc_path) => {
-            if is_system_mode {
+            if enable_eravm_extensions {
                 anyhow::bail!("Yul validation cannot be done if EraVM extensions are enabled. Consider compiling without `solc`.")
             }
             let solc_compiler = SolcCompiler::new(solc_path.as_str())?;
@@ -95,7 +95,7 @@ pub fn yul_to_eravm(
     let build = project.compile_to_eravm(
         optimizer_settings,
         llvm_options,
-        is_system_mode,
+        enable_eravm_extensions,
         include_metadata_hash,
         zkevm_assembly::RunningVmEncodingMode::Production,
         debug_config,
@@ -112,7 +112,7 @@ pub fn yul_to_evm(
     libraries: Vec<String>,
     solc_path: Option<String>,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
+    llvm_options: Vec<String>,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
@@ -150,8 +150,7 @@ pub fn yul_to_evm(
 pub fn llvm_ir_to_eravm(
     paths: &[PathBuf],
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
-    is_system_mode: bool,
+    llvm_options: Vec<String>,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
@@ -160,7 +159,7 @@ pub fn llvm_ir_to_eravm(
     let build = project.compile_to_eravm(
         optimizer_settings,
         llvm_options,
-        is_system_mode,
+        false,
         include_metadata_hash,
         zkevm_assembly::RunningVmEncodingMode::Production,
         debug_config,
@@ -175,7 +174,7 @@ pub fn llvm_ir_to_eravm(
 pub fn llvm_ir_to_evm(
     paths: &[PathBuf],
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
+    llvm_options: Vec<String>,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
@@ -196,7 +195,7 @@ pub fn llvm_ir_to_evm(
 ///
 pub fn eravm_assembly(
     paths: &[PathBuf],
-    llvm_options: &[&str],
+    llvm_options: Vec<String>,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
@@ -224,16 +223,16 @@ pub fn standard_output_eravm(
     solc_compiler: &SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
-    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
     force_evmla: bool,
-    is_system_mode: bool,
+    enable_eravm_extensions: bool,
     include_metadata_hash: bool,
     use_literal_content: bool,
     base_path: Option<String>,
     include_paths: Vec<String>,
     allow_paths: Option<String>,
     remappings: Option<BTreeSet<String>>,
+    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
+    llvm_options: Vec<String>,
     suppressed_warnings: Option<Vec<Warning>>,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
@@ -252,14 +251,16 @@ pub fn standard_output_eravm(
             None,
             &solc_version.default,
             optimizer_settings.is_fallback_to_size_enabled(),
-            optimizer_settings.is_system_request_memoization_disabled(),
-            optimizer_settings.jump_table_density_threshold(),
         ),
         Some(SolcStandardJsonInputSettingsMetadata::new(
             era_compiler_llvm_context::EraVMMetadataHash::None,
             use_literal_content,
         )),
-        solc_pipeline == SolcPipeline::Yul,
+        force_evmla,
+        false,
+        enable_eravm_extensions,
+        false,
+        llvm_options.clone(),
         suppressed_warnings,
     )?;
     let sources = solc_input.sources()?;
@@ -300,7 +301,7 @@ pub fn standard_output_eravm(
     let build = project.compile_to_eravm(
         optimizer_settings,
         llvm_options,
-        is_system_mode,
+        enable_eravm_extensions,
         include_metadata_hash,
         zkevm_assembly::RunningVmEncodingMode::Production,
         debug_config,
@@ -318,8 +319,6 @@ pub fn standard_output_evm(
     solc_compiler: &SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
-    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
     force_evmla: bool,
     include_metadata_hash: bool,
     use_literal_content: bool,
@@ -327,6 +326,8 @@ pub fn standard_output_evm(
     include_paths: Vec<String>,
     allow_paths: Option<String>,
     remappings: Option<BTreeSet<String>>,
+    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
+    llvm_options: Vec<String>,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
     let solc_version = solc_compiler.version.to_owned();
@@ -344,14 +345,16 @@ pub fn standard_output_evm(
             None,
             &solc_version.default,
             optimizer_settings.is_fallback_to_size_enabled(),
-            optimizer_settings.is_system_request_memoization_disabled(),
-            optimizer_settings.jump_table_density_threshold(),
         ),
         Some(SolcStandardJsonInputSettingsMetadata::new(
             era_compiler_llvm_context::EraVMMetadataHash::None,
             use_literal_content,
         )),
-        solc_pipeline == SolcPipeline::Yul,
+        force_evmla,
+        false,
+        false,
+        false,
+        llvm_options.clone(),
         None,
     )?;
     let sources = solc_input.sources()?;
@@ -404,11 +407,10 @@ pub fn standard_output_evm(
 ///
 pub fn standard_json_eravm(
     solc_compiler: Option<&SolcCompiler>,
-    llvm_options: &[&str],
-    json_path: Option<PathBuf>,
-    detect_missing_libraries: bool,
     force_evmla: bool,
-    is_system_mode: bool,
+    enable_eravm_extensions: bool,
+    detect_missing_libraries: bool,
+    json_path: Option<PathBuf>,
     base_path: Option<String>,
     include_paths: Vec<String>,
     allow_paths: Option<String>,
@@ -419,15 +421,29 @@ pub fn standard_json_eravm(
     let mut solc_input = SolcStandardJsonInput::try_from_reader(json_path.as_deref())?;
     let language = solc_input.language;
     let sources = solc_input.sources()?;
+    let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
+
     let optimizer_settings =
         era_compiler_llvm_context::OptimizerSettings::try_from(&solc_input.settings.optimizer)?;
+    let llvm_options = solc_input.settings.llvm_options.take().unwrap_or_default();
+
+    let force_evmla = solc_input.settings.force_evmla.unwrap_or_default() || force_evmla;
+    let enable_eravm_extensions = solc_input
+        .settings
+        .enable_eravm_extensions
+        .unwrap_or_default()
+        || enable_eravm_extensions;
+    let detect_missing_libraries = solc_input
+        .settings
+        .detect_missing_libraries
+        .unwrap_or_default()
+        || detect_missing_libraries;
     let include_metadata_hash = match solc_input.settings.metadata {
         Some(ref metadata) => {
             metadata.bytecode_hash != Some(era_compiler_llvm_context::EraVMMetadataHash::None)
         }
         None => true,
     };
-    let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
 
     let (mut solc_output, solc_version, project) = match (language, solc_compiler) {
         (SolcStandardJsonInputLanguage::Solidity, Some(solc_compiler)) => {
@@ -516,7 +532,7 @@ pub fn standard_json_eravm(
         let build = project.compile_to_eravm(
             optimizer_settings,
             llvm_options,
-            is_system_mode,
+            enable_eravm_extensions,
             include_metadata_hash,
             zkevm_assembly::RunningVmEncodingMode::Production,
             debug_config,
@@ -532,9 +548,8 @@ pub fn standard_json_eravm(
 ///
 pub fn standard_json_evm(
     solc_compiler: Option<&SolcCompiler>,
-    llvm_options: &[&str],
-    json_path: Option<PathBuf>,
     force_evmla: bool,
+    json_path: Option<PathBuf>,
     base_path: Option<String>,
     include_paths: Vec<String>,
     allow_paths: Option<String>,
@@ -545,15 +560,18 @@ pub fn standard_json_evm(
     let mut solc_input = SolcStandardJsonInput::try_from_reader(json_path.as_deref())?;
     let language = solc_input.language;
     let sources = solc_input.sources()?;
+    let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
+
     let optimizer_settings =
         era_compiler_llvm_context::OptimizerSettings::try_from(&solc_input.settings.optimizer)?;
+    let llvm_options = solc_input.settings.llvm_options.take().unwrap_or_default();
+
     let include_metadata_hash = match solc_input.settings.metadata {
         Some(ref metadata) => {
             metadata.bytecode_hash != Some(era_compiler_llvm_context::EraVMMetadataHash::None)
         }
         None => true,
     };
-    let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
 
     let (mut solc_output, solc_version, project) = match (language, solc_compiler) {
         (SolcStandardJsonInputLanguage::Solidity, Some(solc_compiler)) => {
@@ -647,20 +665,20 @@ pub fn combined_json_eravm(
     solc_compiler: &SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
-    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
     force_evmla: bool,
-    is_system_mode: bool,
+    enable_eravm_extensions: bool,
     include_metadata_hash: bool,
     use_literal_content: bool,
     base_path: Option<String>,
     include_paths: Vec<String>,
     allow_paths: Option<String>,
     remappings: Option<BTreeSet<String>>,
-    suppressed_warnings: Option<Vec<Warning>>,
-    debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     output_directory: Option<PathBuf>,
     overwrite: bool,
+    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
+    llvm_options: Vec<String>,
+    suppressed_warnings: Option<Vec<Warning>>,
+    debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<()> {
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
 
@@ -670,16 +688,16 @@ pub fn combined_json_eravm(
         solc_compiler,
         evm_version,
         solc_optimizer_enabled,
-        optimizer_settings,
-        llvm_options,
         force_evmla,
-        is_system_mode,
+        enable_eravm_extensions,
         include_metadata_hash,
         use_literal_content,
         base_path,
         include_paths,
         allow_paths,
         remappings,
+        optimizer_settings,
+        llvm_options,
         suppressed_warnings,
         debug_config,
     )?;
@@ -714,8 +732,6 @@ pub fn combined_json_evm(
     solc_compiler: &SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
-    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
-    llvm_options: &[&str],
     force_evmla: bool,
     include_metadata_hash: bool,
     use_literal_content: bool,
@@ -723,9 +739,11 @@ pub fn combined_json_evm(
     include_paths: Vec<String>,
     allow_paths: Option<String>,
     remappings: Option<BTreeSet<String>>,
-    debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     output_directory: Option<PathBuf>,
     overwrite: bool,
+    optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
+    llvm_options: Vec<String>,
+    debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<()> {
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
 
@@ -735,8 +753,6 @@ pub fn combined_json_evm(
         solc_compiler,
         evm_version,
         solc_optimizer_enabled,
-        optimizer_settings,
-        llvm_options,
         force_evmla,
         include_metadata_hash,
         use_literal_content,
@@ -744,6 +760,8 @@ pub fn combined_json_evm(
         include_paths,
         allow_paths,
         remappings,
+        optimizer_settings,
+        llvm_options,
         debug_config,
     )?;