+++ src/process/input_eravm.rs
@@ -21,8 +21,8 @@ pub struct Input<'a> {
     pub contract: Cow<'a, Contract>,
     /// The project representation.
     pub project: Cow<'a, Project>,
-    /// The system mode flag.
-    pub is_system_mode: bool,
+    /// Whether to enable EraVM extensions.
+    pub enable_eravm_extensions: bool,
     /// Whether to append the metadata hash.
     pub include_metadata_hash: bool,
     /// Enables the test bytecode encoding.
@@ -42,7 +42,7 @@ impl<'a> Input<'a> {
     pub fn new(
         contract: Cow<'a, Contract>,
         project: Cow<'a, Project>,
-        is_system_mode: bool,
+        enable_eravm_extensions: bool,
         include_metadata_hash: bool,
         enable_test_encoding: bool,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
@@ -52,7 +52,7 @@ impl<'a> Input<'a> {
         Self {
             contract,
             project,
-            is_system_mode,
+            enable_eravm_extensions,
             include_metadata_hash,
             enable_test_encoding,
             optimizer_settings,