+++ src/solc/standard_json/input/mod.rs
@@ -76,7 +76,11 @@ impl Input {
         output_selection: SolcStandardJsonInputSettingsSelection,
         optimizer: SolcStandardJsonInputSettingsOptimizer,
         metadata: Option<SolcStandardJsonInputSettingsMetadata>,
+        force_evmla: bool,
         via_ir: bool,
+        enable_eravm_extensions: bool,
+        detect_missing_libraries: bool,
+        llvm_options: Vec<String>,
         suppressed_warnings: Option<Vec<Warning>>,
     ) -> anyhow::Result<Self> {
         let mut paths: BTreeSet<PathBuf> = paths.iter().cloned().collect();
@@ -102,8 +106,12 @@ impl Input {
                 libraries,
                 remappings,
                 output_selection,
+                force_evmla,
                 via_ir,
+                enable_eravm_extensions,
+                detect_missing_libraries,
                 optimizer,
+                llvm_options,
                 metadata,
             ),
             suppressed_warnings,
@@ -121,7 +129,11 @@ impl Input {
         output_selection: SolcStandardJsonInputSettingsSelection,
         optimizer: SolcStandardJsonInputSettingsOptimizer,
         metadata: Option<SolcStandardJsonInputSettingsMetadata>,
+        force_evmla: bool,
         via_ir: bool,
+        enable_eravm_extensions: bool,
+        detect_missing_libraries: bool,
+        llvm_options: Vec<String>,
         suppressed_warnings: Option<Vec<Warning>>,
     ) -> anyhow::Result<Self> {
         let sources = sources
@@ -137,8 +149,12 @@ impl Input {
                 libraries,
                 remappings,
                 output_selection,
+                force_evmla,
                 via_ir,
+                enable_eravm_extensions,
+                detect_missing_libraries,
                 optimizer,
+                llvm_options,
                 metadata,
             ),
             suppressed_warnings,
@@ -152,6 +168,7 @@ impl Input {
         sources: BTreeMap<String, String>,
         libraries: BTreeMap<String, BTreeMap<String, String>>,
         optimizer: SolcStandardJsonInputSettingsOptimizer,
+        llvm_options: Vec<String>,
     ) -> Self {
         let sources = sources
             .into_iter()
@@ -168,7 +185,11 @@ impl Input {
                 None,
                 output_selection,
                 false,
+                false,
+                false,
+                false,
                 optimizer,
+                llvm_options,
                 None,
             ),
             suppressed_warnings: None,
@@ -182,6 +203,7 @@ impl Input {
         paths: &[PathBuf],
         libraries: BTreeMap<String, BTreeMap<String, String>>,
         optimizer: SolcStandardJsonInputSettingsOptimizer,
+        llvm_options: Vec<String>,
     ) -> Self {
         let sources = paths
             .iter()
@@ -203,7 +225,11 @@ impl Input {
                 None,
                 output_selection,
                 false,
+                false,
+                false,
+                false,
                 optimizer,
+                llvm_options,
                 None,
             ),
             suppressed_warnings: None,