+++ src/solc/standard_json/input/settings/optimizer/mod.rs
@@ -29,12 +29,6 @@ pub struct Optimizer {
     /// Whether to try to recompile with -Oz if the bytecode is too large.
     #[serde(skip_serializing)]
     pub fallback_to_optimizing_for_size: Option<bool>,
-    /// Whether to disable the system request memoization.
-    #[serde(skip_serializing)]
-    pub disable_system_request_memoization: Option<bool>,
-    /// Set the jump table density threshold.
-    #[serde(skip_serializing)]
-    pub jump_table_density_threshold: Option<u32>,
 }
 
 impl Optimizer {
@@ -46,8 +40,6 @@ impl Optimizer {
         mode: Option<char>,
         version: &semver::Version,
         fallback_to_optimizing_for_size: bool,
-        disable_system_request_memoization: bool,
-        jump_table_density_threshold: Option<u32>,
     ) -> Self {
         Self {
             enabled,
@@ -58,8 +50,6 @@ impl Optimizer {
                 None
             },
             fallback_to_optimizing_for_size: Some(fallback_to_optimizing_for_size),
-            disable_system_request_memoization: Some(disable_system_request_memoization),
-            jump_table_density_threshold,
         }
     }
 
@@ -67,14 +57,7 @@ impl Optimizer {
     /// A shortcut constructor for Yul validation.
     ///
     pub fn new_yul_validation() -> Self {
-        Self::new(
-            true,
-            None,
-            &SolcCompiler::LAST_SUPPORTED_VERSION,
-            false,
-            false,
-            None,
-        )
+        Self::new(true, None, &SolcCompiler::LAST_SUPPORTED_VERSION, false)
     }
 
     ///
@@ -99,9 +82,6 @@ impl TryFrom<&Optimizer> for era_compiler_llvm_context::OptimizerSettings {
         };
         if value.fallback_to_optimizing_for_size.unwrap_or_default() {
             result.enable_fallback_to_size();
-        }
-        if value.disable_system_request_memoization.unwrap_or_default() {
-            result.disable_system_request_memoization();
         }
         Ok(result)
     }